# swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Pages API
  description: >
    ## Summary

    The Pages API is domain service for Pages

    ## Uploading an Asset

    1. A client requests to upload an asset to  `/documents/{documentId}/assets/request-upload` enpoint and it returns urls for upload the assets

    2. The client uploads the asset by making a binary `POST` request to the url with the appropriate `Content-Type` header (e.g. image/jpeg).

    3. The client confirms success upload by calling `/documents/{documentId}/assets/finish-upload` endpoint and it returns asset urls for given ids.
host: pages-api-svc
basePath: /v1
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
responses:
  '400':
    description: 'Request was unable to parse, bad JSON or query arguments'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '403':
    description: Permissions denied
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '404':
    description: Resource not found
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '422':
    description: 'Invalid request, missing required parameters'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
  '500':
    description: 'Server encountered a critical error, but the request appears valid'
    schema:
      type: object
      properties:
        message:
          description: |
            Human-readable message for the error, should provide remediation
            instructions
          type: string
paths:
...
  '/teams/{teamId}/documents':
    get:
      tags:
        - Documents
      summary: List of team documents
      description: >
        Returns a list of Document objects based on given `teamId`
      parameters:
        - $ref: '#/parameters/teamId'
        - $ref: '#/parameters/userIdInQuery'
        - $ref: '#/parameters/includeThumbnails'
        - $ref: '#/parameters/isArchived'
        - $ref: '#/parameters/documentIds'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              documents:
                type: array
                items:
                  $ref: '#/definitions/Document'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
  '/teams/{teamId}/users/{userId}/documents':
    get:
      tags:
        - Documents
      summary: List of user documents in given team
      description: >
        Returns a list of Document objects based on given `teamId` and `userId`
      parameters:
        - $ref: '#/parameters/teamId'
        - $ref: '#/parameters/userId'
        - $ref: '#/parameters/userIdInQuery'
        - $ref: '#/parameters/includeThumbnails'
        - $ref: '#/parameters/isArchived'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              documents:
                type: array
                items:
                  $ref: '#/definitions/Document'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
...
  '/private/teams/{teamId}/documents':
    get:
      tags:
        - Private
      summary: List of team documents
      description: >
        Returns a list of Document objects based on given `teamId`
      parameters:
        - $ref: '#/parameters/teamId'
        - $ref: '#/parameters/includeThumbnails'
        - $ref: '#/parameters/isArchived'
        - $ref: '#/parameters/documentIds'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              documents:
                type: array
                items:
                  $ref: '#/definitions/Document'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
        '404':
          $ref: '#/responses/404'
        '422':
          $ref: '#/responses/422'
        '500':
          $ref: '#/responses/500'
...
definitions:
...
  DocumentMembership:
    type: object
    properties:
      userId:
        type: number
        example: 1
...
  Document:
    type: object
    required:
    - id
    - title
    - ownerId
    - teamId
    - url
    - memberships
    - createdAt
    - updatedAt
    - isArchived
    properties:
      id:
        type: string
        format: uuid
        example: 'ccd533e7-d917-4601-b708-9bfea3e26f04'
      title:
        type: string
        example: 'Untitled'
      ownerId:
        type: integer
        description: Owner user id
        example: 1
      teamId:
        type: string
        example: 'cjcjeoi2w0000rn35c23q98ou'
      url:
        type: string
        example: '/rhombus/Untitled-eXNEfNKEHjaDSqU4gZUv8M'
      memberships:
        type: array
        items:
          $ref: '#/definitions/DocumentMembership'
      createdAt:
        type: string
        format: date-time
        example: '2018-02-28T21:39:18.000Z'
      updatedAt:
        type: string
        format: date-time
        example: '2018-02-28T21:39:18.000Z'
      archivedAt:
        default: null
        type: string
        format: date-time
        example: '2018-02-28T21:39:18.000Z'
      isArchived:
        type: boolean
        example: false
...
parameters:
  userId:
    name: userId
    in: path
    description: The ID of the User
    type: integer
    required: true
  userIdInQuery:
    name: user_id
    in: query
    description: The ID of the User
    type: integer
    required: true
...
  teamId:
    name: teamId
    in: path
    description: The ID of the Team
    type: string
    required: true
...
  documentIds:
    name: documentIds
    description: >
      Optional. A list of document ID's.

      Could be also list of comma separated values
    in: query
    type: array
    items:
      type: string
...
  includeThumbnails:
    name: includeThumbnails
    default: false
    description: >
      Optional. Include thumbnails in response.
    in: query
    type: boolean
...
  isArchived:
    name: isArchived
    default: false
    description: >
      Optional. Return archived documents
    in: query
    type: boolean